+++ packages/graphql-docs-generator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/graphql-docs-generator@4.1.0...@aws-amplify/graphql-docs-generator@4.2.0) (2023-09-12)
+
+### Features
+
+- graphql generator ([#690](https://github.com/aws-amplify/amplify-codegen/issues/690)) ([dbe12ab](https://github.com/aws-amplify/amplify-codegen/commit/dbe12abbbcd307bec1c15f95251f023d0f0fbf10))
+- graphql-generator package ([#677](https://github.com/aws-amplify/amplify-codegen/issues/677)) ([6348627](https://github.com/aws-amplify/amplify-codegen/commit/634862793cb5aebb284f27a70f0ef07d6fd85561)), closes [#669](https://github.com/aws-amplify/amplify-codegen/issues/669) [#670](https://github.com/aws-amplify/amplify-codegen/issues/670) [#671](https://github.com/aws-amplify/amplify-codegen/issues/671) [#676](https://github.com/aws-amplify/amplify-codegen/issues/676) [#680](https://github.com/aws-amplify/amplify-codegen/issues/680)
+
+### Reverts
+
+- Revert "feat: graphql-generator package (#677)" (#687) ([d7a84d6](https://github.com/aws-amplify/amplify-codegen/commit/d7a84d62b5a1e686d4e5e5be61e12fb410378685)), closes [#677](https://github.com/aws-amplify/amplify-codegen/issues/677) [#687](https://github.com/aws-amplify/amplify-codegen/issues/687)
+
 # [4.1.0](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/graphql-docs-generator@4.0.5...@aws-amplify/graphql-docs-generator@4.1.0) (2023-08-25)
 
 ### Features